+++ cipher/viper_test.go
@@ -1,6 +1,8 @@
 package cipher
 
 import (
+	"fmt"
+	"github.com/Comcast/webpa-common/logging"
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/assert"
 	"os"
@@ -23,9 +25,9 @@ func TestViper(t *testing.T) {
 	options, err := FromViper(v)
 	assert.NoError(err)
 
-	encrypter, err := options.LoadEncrypt()
+	encrypter, err := options.GetEncrypter(logging.NewTestLogger(nil, t))
 	assert.NoError(err)
-	assert.NotEmpty(options)
+	assert.NotNil(encrypter)
 
 	msg := "hello"
 	data, _, err := encrypter.EncryptMessage([]byte(msg))
@@ -49,7 +51,7 @@ func TestNOOPViper(t *testing.T) {
 	options, err := FromViper(v)
 	assert.NoError(err)
 
-	encrypter, err := options.LoadEncrypt()
+	encrypter, err := options.GetEncrypter(logging.NewTestLogger(nil, t))
 
 	msg := "hello"
 	data, _, err := encrypter.EncryptMessage([]byte(msg))
@@ -72,7 +74,7 @@ func TestBoxBothSides(t *testing.T) {
 	options, err := FromViper(vSend)
 	assert.NoError(err)
 
-	encrypter, err := options.LoadEncrypt()
+	encrypter, err := options.GetEncrypter(logging.NewTestLogger(nil, t))
 	assert.NoError(err)
 
 	vRec := viper.New()
@@ -86,16 +88,56 @@ func TestBoxBothSides(t *testing.T) {
 	options, err = FromViper(vRec)
 	assert.NoError(err)
 
-	decrypter, err := options.LoadDecrypt()
+	decrypters := PopulateCiphers(options, logging.NewTestLogger(nil, t))
+
 	assert.NoError(err)
 
 	msg := []byte("hello")
 	data, nonce, err := encrypter.EncryptMessage(msg)
 	assert.NoError(err)
 
-	decodedMSG, err := decrypter.DecryptMessage(data, nonce)
+	if decrypter, ok := decrypters.Get(encrypter.GetAlgorithm(), encrypter.GetKID()); ok {
+		decodedMSG, err := decrypter.DecryptMessage(data, nonce)
+		assert.NoError(err)
+
+		assert.Equal(msg, decodedMSG)
+	} else {
+		assert.Fail("failed to get decrypter with kid")
+	}
+}
+
+func TestGetDecrypterErr(t *testing.T) {
+	assert := assert.New(t)
+
+	vSend := viper.New()
+	path, err := os.Getwd()
+	assert.NoError(err)
+	vSend.AddConfigPath(path)
+	vSend.SetConfigName("boxRecipient")
+	if err := vSend.ReadInConfig(); err != nil {
+		t.Fatalf("%s\n", err)
+	}
+
+	options, err := FromViper(vSend)
 	assert.NoError(err)
 
-	assert.Equal(msg, decodedMSG)
+	decrypters := PopulateCiphers(options, logging.NewTestLogger(nil, t))
+	fmt.Printf("%#v\n", decrypters)
 
+	decrypter, ok := decrypters.Get(Box, "test")
+	assert.True(ok)
+	assert.NotNil(decrypter)
+
+	decrypter, ok = decrypters.Get(None, "none")
+	assert.True(ok)
+	assert.NotNil(decrypter)
+
+	// negative test
+	decrypter, ok = decrypters.Get(None, "neato")
+	assert.False(ok)
+	assert.Nil(decrypter)
+
+	decrypter, ok = decrypters.Get(RSAAsymmetric, "testing")
+	assert.False(ok)
+	assert.Nil(decrypter)
 }