+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added kid and alg to `db` package
+- Updated loading cipher options
 
 ## [v0.3.3]
 - Fix box loader in `cipher` package